Spawning Tribler subprocess now
