--- TriblerGUI/scripts/__init__.py
@@ -1,3 +1,0 @@
-"""
-This package contains various utility scripts.
-"""
